requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
google-generativeai
scikit-learn
pandas
numpy<2.0
scikit-surprise
apscheduler
certifi
pydantic